Add cluster tree mixin and update hierarchical clusterers
