+++ lib/ai4r/clusterers/average_linkage.rb
@@ -10,6 +10,7 @@
 
 require_relative '../data/data_set'
 require_relative '../clusterers/single_linkage'
+require_relative '../clusterers/cluster_tree'
 
 module Ai4r
   module Clusterers
@@ -26,6 +27,8 @@ module Ai4r
     #
     #   D(cx, (ci U cj) = (D(cx, ci) + D(cx, cj)) / 2
     class AverageLinkage < SingleLinkage
+
+      include ClusterTree
 
       parameters_info :distance_function =>
           "Custom implementation of distance function. " +
