+++ lib/ai4r/clusterers/centroid_linkage.rb
@@ -10,6 +10,7 @@
 
 require_relative '../data/data_set'
 require_relative '../clusterers/single_linkage'
+require_relative '../clusterers/cluster_tree'
 
 module Ai4r
   module Clusterers
@@ -29,6 +30,8 @@ module Ai4r
     #                       (nj/(ni+nj))*D(cx, cj) -
     #                       (ni*nj/(ni+nj)^2)*D(ci, cj)
     class CentroidLinkage < SingleLinkage
+
+      include ClusterTree
 
     parameters_info :distance_function =>
           "Custom implementation of distance function. " +
