+++ lib/ai4r/clusterers/ward_linkage.rb
@@ -10,6 +10,7 @@
 
 require_relative '../data/data_set'
 require_relative '../clusterers/single_linkage'
+require_relative '../clusterers/cluster_tree'
 
 module Ai4r
   module Clusterers
@@ -26,6 +27,8 @@ module Ai4r
     #                       (nj/(ni+nj+nx))*D(cx, cj) -
     #                       (nx/(ni+nj)^2)*D(ci, cj)
     class WardLinkage < SingleLinkage
+
+      include ClusterTree
 
     parameters_info :distance_function =>
           "Custom implementation of distance function. " +
