+++ lib/ai4r/clusterers/weighted_average_linkage.rb
@@ -10,6 +10,7 @@
 
 require_relative '../data/data_set'
 require_relative '../clusterers/single_linkage'
+require_relative '../clusterers/cluster_tree'
 
 module Ai4r
   module Clusterers
@@ -25,6 +26,8 @@ module Ai4r
     #
     #   D(cx, (ci U cj)) =  ( ni * D(cx, ci) + nj * D(cx, cj)) / (ni + nj)
     class WeightedAverageLinkage < SingleLinkage
+
+      include ClusterTree
 
     parameters_info :distance_function =>
           "Custom implementation of distance function. " +
